ACTman - Actigraphy Manager

ACTman manages actigraphy data whilst offering pre-processing and analyses options. This initial version supports the 'Actiwatch 2' and 'MotionWatch 8' actigraphy devices, whilst allowing for both sleep and circadian rhythm analyses.

ACTman(workdir = "C:/Bibliotheek/Studie/PhD/Publishing/ACTman/R-part/mydata2",
  sleepdatadir = "C:/Bibliotheek/Studie/PhD/Publishing/ACTman/R-part/Actogram & Sleep analysis",
  myACTdevice = "Actiwatch2", iwantsleepanalysis = FALSE,
  plotactogram = FALSE, selectperiod = FALSE, startperiod, daysperiod,
  movingwindow = FALSE, movingwindow.size = 14, lengthcheck = FALSE,
  na_omit = TRUE)

Arguments

workdir
The working directory of the script.

sleepdatadir
An optional vector specifying the directory for actogram and sleep analysis data.

myACTdevice
Name of the input device used. Should be either 'Actiwatch2' or 'MW8'.

iwantsleepanalysis
Boolean value indicating whether sleep analysis should be performed.

plotactogram
Boolean value indicating whether an actogram has to be plotted.

selectperiod
Boolean value indicating whether a specific period has to be selected.

startperiod
An optional vector specifying single or multiple period starts. Should be in the format "2016-10-03 00:00:00".

daysperiod
An optional vector specifying the length in days of the period.

movingwindow
Boolean value indicating whether a moving window should be utilised.

movingwindow.size
An optional vector specifying the length in days of the moving window. Default is 14 days.

lengthcheck
Boolean value indicating whether data recorded after 14 days should be included.

na_omit
Boolean value indicating whether NA's should be omitted.

Value

nothing

Examples

# NOT RUN {
View(ACTman::ACTman(workdir = "C:/Bibliotheek/Studie/PhD/Publishing/ACTman/R-part/mydata2",
                    myACTdevice = "Actiwatch2",
                    iwantsleepanalysis = FALSE,
                    plotactogram = FALSE,
                    selectperiod = FALSE,
                    startperiod = "2016-10-03 00:00:00",
                    daysperiod = 14, movingwindow = FALSE, lengthcheck = TRUE))
# }